🚚 Organize Category presentation files
Files updated in more packages in order to better separate the concerns
and make the code easier to navigate.
