+++ features/category/src/main/java/com/escodro/category/presentation/bottomsheet/CategoryAddViewModel.kt
@@ -1,4 +1,4 @@
-package com.escodro.category.presentation
+package com.escodro.category.presentation.bottomsheet
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
